Setting up the init python files to allow for easy imports and one step database connections.
