+++ .gitignore
@@ -1,2 +1,3 @@
 config.py
 index.json
+/mongodb/config.py
