+++ mongodb/__init__.py
@@ -1,0 +1,12 @@
+from pymongo import MongoClient
+import config
+
+# MongoDB setup
+# Connect to the MongoDB server on DigitalOcean
+client = MongoClient(config.mongo_db_connection)
+# Create (or use existing) database
+db = client['conversation_db']
+
+__all__ = [
+    db
+]
