+++ mongodb/insert_conversations.py
@@ -1,6 +1,4 @@
-from pymongo import MongoClient
-import json
-import config
+from mongodb import db
 from datetime import datetime, timezone
 import uuid
 
@@ -20,21 +18,11 @@ def generate_conversation_json(conversation_id, user_name, messages):
 
 def conversation_to_mongo(user_name, messages):
 
-    # Connect to the MongoDB server on DigitalOcean
-    client = MongoClient(config.mongo_db_connection)
-
-    # Create (or use existing) database
-    db = client['conversation_db']
-
     # Create (or use existing) collection
     collection = db['conversations']
 
     # make a random UUID
     conversation_id = str(uuid.uuid4())
-
-    #generate JSON
-    # json = generate_conversation_json(conversation_id, user_name, messages)
-    # print(json)
 
     # Insert the conversation into the collection
     collection.insert_one(generate_conversation_json(conversation_id, user_name, messages))
@@ -43,22 +31,4 @@ def conversation_to_mongo(user_name, messages):
     collection.create_index(user_name)
 
     print("Conversation inserted and index created.")
-
-
-
-# Legacy
-# Opening JSON file
-# f = open('./utils/sample_conversation.json')
- 
-# returns JSON object as 
-# a dictionary
-# conversation = json.load(f)
-
-# # Sample JSON conversation data
-# conversation_data = 
-
-# # Load JSON data
-# conversation = json.loads(utils/sample)
-
-
