+++ requirement.txt
@@ -1,0 +1,3 @@
+pymongo
+langchain
+openai
